Move `KeyStatusListener` into `ServiceEndpoint`
